requirements.txt
mysql-connector-python~=9.1.0
requests~=2.32.3